exception handling - delete

+++ src/main/java/com/rozangelapeixoto/course/services/UserService.java
@@ -2,10 +2,11 @@ package com.rozangelapeixoto.course.services;
 
 import com.rozangelapeixoto.course.entities.User;
 import com.rozangelapeixoto.course.repositories.UserRepository;
+import com.rozangelapeixoto.course.services.exceptions.DatabaseExpection;
 import com.rozangelapeixoto.course.services.exceptions.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
-
 
 import java.util.List;
 import java.util.Optional;
@@ -29,7 +30,14 @@ public class UserService {
     }
 
     public void delete(Long id){
-        repository.deleteById(id);
+        if(!repository.existsById(id)){
+            throw new ResourceNotFoundException(id);
+        }
+        try {
+            repository.deleteById(id);
+        }catch(DataIntegrityViolationException e ){
+            throw new DatabaseExpection(e.getMessage());
+        }
     }
 
     public User update(Long id, User obj){
